+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+_No unreleased changes_
+
+## [0.6.0] - 2020-08-25
 - Added resize mode options
 - Added output size options for emotes (28px, 56px, 112px) and badges (18px, 36px, 72px)
 - Added an image icon to drag area
 - Added status messages
-- Changed image resizing library to Jimp.
+- Changed image resizing library to Jimp
 
 ## [0.5.0] - 2020-08-23
 - Fixed help button
@@ -30,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Initial Release
 
+[0.6.0]: https://github.com/VariXx/emote-resizer/tree/v0.6.0
 [0.5.0]: https://github.com/VariXx/emote-resizer/tree/v0.5.0
 [0.4.0]: https://github.com/VariXx/emote-resizer/tree/v0.4.0
 [0.3.0]: https://github.com/VariXx/emote-resizer/tree/v0.3.0